Fix memory leak
The variables were passed by value for the swap function.

+++ point.hpp
@@ -113,7 +113,7 @@ public:
 };
 
 
-typedef double point_t;
+typedef int32_t point_t;
 typedef point<point_t> pt;
 typedef std::pair<pt,pt> edge;
